requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.1.0
blinker==1.6.2
cairocffi==1.6.0
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
chess==1.9.4
click==8.1.6
cssselect2==0.7.0
defusedxml==0.7.1
discord.py==2.3.1
Flask==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
tinycss2==1.2.1
topggpy==1.4.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uWSGI==2.0.21
webencodings==0.5.1
Werkzeug==2.3.6
yarl==1.9.2
